split up backend fetch into multiple files
